move bootstrap publishing code to its own module
